Added 2 new options for BACKLIGHT and DIMMER: 'trans_fixed'.
New options let users set a fixed smooth transitions duration; the new options are off by default for now,
but are most probably going to be enabled by default as I found that having a fixed transition duration (eg: 1000ms)
for any backlight-related update enables a much better (and smooth) user experience.

In the event that the new options will be enabled by default, it will be obviously stated in the release notes.
For now, I don't really want to break existing configurations (as 'trans_step' and 'trans_timeout' become useless when 'trans_fixed' is enabled.

Refs #233

Signed-off-by: Federico Di Pierro <nierro92@gmail.com>